Determine the terminal background color just once, at startup (#1504)
Currently we may determine it twice (once for the logo, once for the
theme), and the theme determining can happen at multiple different
times depending on the code path.

This weirds me out, so I'm changing it.

I was sort of hoping that this would fix the problem where sometimes I
would see the reponse from the terminal make it to my screen instead
of getting captured by the library, and it might have. I haven't see
it with this patch applied.
But I also don't really think it *should* fix it, so, we'll see.