Allow doing `edgedb cli upgrade` for a CLI anywhere under $HOME (#1501)
The goal here is to allow doing it from our language-binding-managed
CLI installations.

In order to make this work, we need to pass an `--installation-path`
to the downloaded new CLI version. This was added in #1499, which has
been released on nightly, testing, and stable channels. I tested
upgrading to all of those channels.

A side effect of doing this is that this removes support for migration
from the "old" binary path of `.edgedb/bin/`. That migration happened
in 2021, though, I think, so I'm not super worried.