Usually eliminate ANSI spew on slow ssh connections (#1522)
At @mmastrac's suggestion, I made the library that queries the
terminal use a "fence" technique where we send a universally-supported
query after the background color query, so that we can detect when the
BG color query isn't handled.

This lets us turn up the timeout quite high (we *maybe* could
eliminate it, but that makes me a little nervous).

I tested on a slow SSH connection using gnome-terminal (which does
support querying BG color) and pterm (which does not).

See:
 * https://github.com/msullivan/xterm-query/pull/1
 * https://github.com/msullivan/terminal-light/pull/1
 * https://github.com/msullivan/terminal-light/pull/2

Fixes #1520.