Support branching commands on cloud without a project
Branching commands all need to know the "current branch".
Current branch is either inferred from the project or
read from credentials.json.

This is a problem for cloud instances, becuase they do not
have a credentials.json so before this PR, we only supported
branching commands on cloud if there was an associated project.

This PR makes branching commands use --branch arg to override
current branch, which then removes the need for instances to have
credentials.json.