Support branching commands on cloud without a project
Supersedes #1574 and #1575

Problem:

Branching commands all need to know the "current branch".
Current branch is either inferred from the project or
read from credentials.json.

This is a problem for cloud instances, becuase they do not
have a credentials.json so before this PR, we only supported
branching commands on cloud if there was an associated project.

Solution:

Since a few commands (`list`, `wipe`, `drop`) do not even
need current branch, this PR just allows specifying --instance,
even if that instance is cloud.

We already do that in a few different cases, see comment of
`Context`.

I've also improved error message when trying to set current
branch of a cloud instance.