Local instance backup (#1585)
This is a first pass at support local instance backups.

It requires a running instance (fails if the instance is sleeping), It doesn't support restore or incremental backups, but it will at least perform a local backup.

Further passes will improve the integration with the instance handle (which is a bit too tied with the current backup operation)