Apply clippy::large_futures lint
`await` macro desugars into a state machine that holds all local
variables at each `.await` point a function. This state machine is
stored on the stack, which might get very large.

On Windows we were hitting stack overflow problems in latest attempts
to test hooks and watch command.

Solution is to wrap futures with `Box::pin`, which moves the state
machine to the heap.

Fortunately, there is a clippy lint that finds large futures.

This PR also applies a few other lints in Windows-only code.