Fix squashing of long migration histories (#1610)
When deleting old migration scripts during squash, we first rename
them to `.old.edgeql` and then delete them.  While looping through the
directory contents, if we see an existing `.old.edgeql`, we register
that to be deleted too.

However, we are iterating through the directory as we do this, so it
is possible that we will see a file that was just *renamed* and thus
is already registered to be deleted, and which we will then try to
delete twice.

On linux, this happens if there too many files to fit in one call to
`getdents64` with whatever buffer size libc is using.  To make things
extra fun, I originally struggled to reproduce this because glibc uses
a much larger buffer size than musl does (32k vs 2k).

Fix by just dropping the deletion of existing `.edgeql.old` files. I
don't think they should be there for any reason of ours, and so we
don't want to mess with them?
I don't feel strongly, though; we could also skip duplicates.

Tested manually; I didn't include a test because the long migration
history is a pain.

Fixes #1609.