Implement incremental backup (#1615)
Performing a backup for an instance will now use an "automatic" mode where we perform an incremental backup if the current backup's "generation" is <= 5 (an arbitrary number to avoid incrementals becoming too deeply nested) or if the previous backup is older than a week.

We also terminate and start the instance appropriately for backup and restore to take place.