Initial locking implementation (#1616)
We add a set of locks on instances and projects that control exclusive or shared access. This allows us to prevent failures when gel watch is running.

Most commands take an exclusive lock, but a few use a shared lock which allows them to run under gel project hooks. I might have missed some commands but it's easy enough to mark them as taking read-locks.

To simplify the implementation, we add locks into higher-level contexts in the CLI rather than individual commands where possible.