Use dunce to canonicalize paths for a better Windows experience (#1668)
If we end up running the gel command in a drive-based UNC path for some reason (ie: \\?\C:\foo\bar), we may not be able to run hooks because they are executed through cmd.exe, and cmd.exe really, really doesn't like UNC paths.

We use dunce to canonicalize paths before setting the cwd of processes to avoid this issue. This won't fix issues where users are running from a network path w/o a drive (unlikely), but it will ensure that 99% of users get a reasonable experience.