Rewrite the shared-client-tests to use a harness-less test (#1720)
We don't need to use a build script for shared-client-tests. This can just be a harness-less test suite.

Also, move it to new-tests so we can fully remove the old portable windows tests.

I had to remove the portable windows tests in #1719, and this restores it, but as part of new-tests.