Bump toolchain to 1.88 (#1724)
let chains are starting to show up in dependencies downstream - let's bump to 1.88 so we can get ahead of these deps.

override: true was provided on all the dtolnay/rust-toolchain invocations, but that actually does nothing. Removed as part of this PR.

Something changed as of Rust 1.88 where it looks like we need to specify the three-digit build rather than just 1.88. Not entirely sure why.