Watch only needed directories
Closes #1649

Before this PR, `gel watch` would subscribe to fs notifications of all
files in the gel project directory. This is slow and inefficient,
because usually, we are only watching for changes in a few files.
It gets bad when there is also a build directorys with thousends of
files.

This PR replaces usage of globset crate for matching glob patterns
with wax crate, because it has [`.partition()` function][^1].
It splits the glob in a invariant prefix and a variant suffix.
For example, `src/**/*.edgeql` is split into `src` and `**/*.edgeql`.

This allows us to subscribe only to changes in directories that we need.

[^1]: https://docs.rs/wax/latest/wax/struct.Glob.html#method.partition